+++ spec/adhearsion/drb/service_spec.rb
@@ -1,12 +1,16 @@
 require 'spec_helper'
 
-describe Adhearsion::Drb::Plugin::Service do
+describe Adhearsion::Drb::Service do
 
   before :all do
     @host = Adhearsion.config[:adhearsion_drb].host
     @port = Adhearsion.config[:adhearsion_drb].port
     @allow = Adhearsion.config[:adhearsion_drb].acl.allow.dup
     @deny = Adhearsion.config[:adhearsion_drb].acl.deny.dup
+    # Use a random high port to prevent concurrent test runs from getting
+    # Errno::EADDRINUSE
+    # Ruby 1.9.2 version of #rand only allows 1 arg; do some math to keep the range right
+    Adhearsion.config[:adhearsion_drb].port = rand(65535 - 1024) + 1024
   end
 
   after :all do
@@ -19,25 +23,25 @@ describe Adhearsion::Drb::Plugin::Service do
   describe "while creating the acl value" do
 
     it "should return <allow 127.0.0.1> as default value" do
-      Adhearsion::Drb::Plugin::Service.create_acl(Adhearsion.config.adhearsion_drb.acl.allow, Adhearsion.config.adhearsion_drb.acl.deny).should == %w<allow 127.0.0.1>
+      described_class.create_acl(Adhearsion.config.adhearsion_drb.acl.allow, Adhearsion.config.adhearsion_drb.acl.deny).should == %w<allow 127.0.0.1>
     end
 
     it "should return an empty string when no rule is defined" do
       Adhearsion.config.adhearsion_drb.acl.allow = []
       Adhearsion.config.adhearsion_drb.acl.deny = []
-      Adhearsion::Drb::Plugin::Service.create_acl(Adhearsion.config.adhearsion_drb.acl.allow, Adhearsion.config.adhearsion_drb.acl.deny).should == []
+      described_class.create_acl(Adhearsion.config.adhearsion_drb.acl.allow, Adhearsion.config.adhearsion_drb.acl.deny).should == []
     end
 
     it "should return an empty string when allow and deny are nil" do
       Adhearsion.config.adhearsion_drb.acl.allow = nil
       Adhearsion.config.adhearsion_drb.acl.deny = nil
-      Adhearsion::Drb::Plugin::Service.create_acl(Adhearsion.config.adhearsion_drb.acl.allow, Adhearsion.config.adhearsion_drb.acl.deny).should == []
+      described_class.create_acl(Adhearsion.config.adhearsion_drb.acl.allow, Adhearsion.config.adhearsion_drb.acl.deny).should == []
     end
 
     it "should return an empty string when allow and deny are empty" do
       Adhearsion.config.adhearsion_drb.acl.allow = ""
       Adhearsion.config.adhearsion_drb.acl.deny = ""
-      Adhearsion::Drb::Plugin::Service.create_acl(Adhearsion.config.adhearsion_drb.acl.allow, Adhearsion.config.adhearsion_drb.acl.deny).should == []
+      described_class.create_acl(Adhearsion.config.adhearsion_drb.acl.allow, Adhearsion.config.adhearsion_drb.acl.deny).should == []
     end
 
     describe "having configured deny" do
@@ -47,12 +51,12 @@ describe Adhearsion::Drb::Plugin::Service do
 
       it "should return an array with <deny 10.1.*.* deny 10.0.*.*>" do
         Adhearsion.config.adhearsion_drb.acl.deny = %w<10.1.*.* 10.0.*.*>
-        Adhearsion::Drb::Plugin::Service.create_acl(Adhearsion.config.adhearsion_drb.acl.allow, Adhearsion.config.adhearsion_drb.acl.deny).should == %w<deny 10.1.*.* deny 10.0.*.*>
+        described_class.create_acl(Adhearsion.config.adhearsion_drb.acl.allow, Adhearsion.config.adhearsion_drb.acl.deny).should == %w<deny 10.1.*.* deny 10.0.*.*>
       end
 
       it "should return an array with <deny 10.1.*.*>" do
         Adhearsion.config.adhearsion_drb.acl.deny = "10.1.*.*"
-        Adhearsion::Drb::Plugin::Service.create_acl(Adhearsion.config.adhearsion_drb.acl.allow, Adhearsion.config.adhearsion_drb.acl.deny).should == %w<deny 10.1.*.*>
+        described_class.create_acl(Adhearsion.config.adhearsion_drb.acl.allow, Adhearsion.config.adhearsion_drb.acl.deny).should == %w<deny 10.1.*.*>
       end
     end
 
@@ -63,12 +67,12 @@ describe Adhearsion::Drb::Plugin::Service do
 
       it "should return an array with <allow 127.0.0.1 allow 10.0.0.1> when another IP is allowed" do
         Adhearsion.config.adhearsion_drb.acl.allow = %w<127.0.0.1 10.0.0.1>
-        Adhearsion::Drb::Plugin::Service.create_acl(Adhearsion.config.adhearsion_drb.acl.allow, Adhearsion.config.adhearsion_drb.acl.deny).should == %w<allow 127.0.0.1 allow 10.0.0.1>
+        described_class.create_acl(Adhearsion.config.adhearsion_drb.acl.allow, Adhearsion.config.adhearsion_drb.acl.deny).should == %w<allow 127.0.0.1 allow 10.0.0.1>
       end
 
       it "should return an array with <allow 10.0.0.1> when another IP is allowed" do
         Adhearsion.config.adhearsion_drb.acl.allow = "10.0.0.1"
-        Adhearsion::Drb::Plugin::Service.create_acl(Adhearsion.config.adhearsion_drb.acl.allow, Adhearsion.config.adhearsion_drb.acl.deny).should == %w<allow 10.0.0.1>
+        described_class.create_acl(Adhearsion.config.adhearsion_drb.acl.allow, Adhearsion.config.adhearsion_drb.acl.deny).should == %w<allow 10.0.0.1>
       end
     end
 
@@ -77,13 +81,13 @@ describe Adhearsion::Drb::Plugin::Service do
       it "should return an array with <allow 127.0.0.1 allow 10.2.*.* deny 10.1.*.* deny 10.0.*.*>" do
         Adhearsion.config.adhearsion_drb.acl.allow = %w<127.0.0.1 10.2.*.*>
         Adhearsion.config.adhearsion_drb.acl.deny = %w<10.1.*.* 10.0.*.*>
-        Adhearsion::Drb::Plugin::Service.create_acl(Adhearsion.config.adhearsion_drb.acl.allow, Adhearsion.config.adhearsion_drb.acl.deny).should == %w<allow 127.0.0.1 allow 10.2.*.* deny 10.1.*.* deny 10.0.*.*>
+        described_class.create_acl(Adhearsion.config.adhearsion_drb.acl.allow, Adhearsion.config.adhearsion_drb.acl.deny).should == %w<allow 127.0.0.1 allow 10.2.*.* deny 10.1.*.* deny 10.0.*.*>
       end
 
       it "should return an array with <allow 127.0.0.1 deny 10.1.*.*>" do
         Adhearsion.config.adhearsion_drb.acl.allow = "127.0.0.1"
         Adhearsion.config.adhearsion_drb.acl.deny = "10.1.*.*"
-        Adhearsion::Drb::Plugin::Service.create_acl(Adhearsion.config.adhearsion_drb.acl.allow, Adhearsion.config.adhearsion_drb.acl.deny).should == %w<allow 127.0.0.1 deny 10.1.*.*>
+        described_class.create_acl(Adhearsion.config.adhearsion_drb.acl.allow, Adhearsion.config.adhearsion_drb.acl.deny).should == %w<allow 127.0.0.1 deny 10.1.*.*>
       end
     end
   end
@@ -97,22 +101,21 @@ describe Adhearsion::Drb::Plugin::Service do
       def foo
         [3,2,1]
       end
+
+      def halt_drb_directly!
+        DRb.stop_service
+      end
     end
 
-    let :client do
-      DRbObject.new nil, DRb.uri
-    end
+    let(:client) { DRbObject.new nil, DRb.uri }
 
-    before do
-      Adhearsion.config.adhearsion_drb.acl.allow = %q<127.0.0.1>
-      Adhearsion.config.adhearsion_drb.acl.deny = nil
+    before(:all) do
+      Adhearsion.config.adhearsion_drb.acl.allow     = %q<127.0.0.1>
+      Adhearsion.config.adhearsion_drb.acl.deny      = nil
       Adhearsion.config.adhearsion_drb.shared_object = Blah.new
-      
-      Adhearsion::Plugin.init_plugins
-    end
 
-    after do
-      Adhearsion::Drb::Plugin::Service.stop
+      Adhearsion::Drb::Service.user_stopped = false
+      Adhearsion::Drb::Service.start
     end
 
     it "should return normal Ruby data structures properly over DRb" do
@@ -123,6 +126,17 @@ describe Adhearsion::Drb::Plugin::Service do
       lambda { client.interface.bad_interface.should be [3, 2, 1] }.should raise_error NoMethodError
     end
 
-  end
+    it "restarts the server if DRb's thread ends" do
+      client.halt_drb_directly!
+      sleep 2
+      client.foo.should == [3, 2, 1]
+    end
 
+    it "does not start up again if Service.stop is called" do
+      Adhearsion::Drb::Service.should_not_receive :start
+      Adhearsion::Drb::Service.stop
+      sleep 0.10
+      lambda { client.foo }.should raise_error DRb::DRbServerNotFound
+    end
+  end
 end