+++ CHANGELOG.md
@@ -1,2 +1,6 @@
+# 1.1.0
+  * Make the server restart when it goes away
+  * Add travis config
+
 # v1.0.0
   * First release